Split LoRA layer implementations into separate files.
